pyproject.toml
[tool.poetry]
name = "word-translater"
version = "0.1.0"
description = ""
authors = ["pibs01241 <i.bennis.ib@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.6"
openai = "^1.53.0"
django = "^5.1.2"
langfuse = "^2.53.3"
langchain-openai = "^0.2.4"
python-docx = "^1.1.2"
langchain-community = "^0.3.4"
langchain-core = "^0.3.15"
openpyxl = "^3.1.5"
python-pptx = "^1.0.2"
nltk = "^3.9.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
ruff = "^0.7.1"
mypy = "^1.13.0"
ipykernel = "^6.29.5"
pytest = "^8.3.3"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=translating_app.settings --reuse-db -v -s"
python_files = ["tests.py", "test_*.py"]
testpaths = ["translating_app", "src"]
junit_family = "xunit2"
filterwarnings = ["ignore::PendingDeprecationWarning"]
